requirements.txt
astroid==2.3.3
autopep8==1.4.4
Click==7.0
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.18
pep8==1.7.1
protobuf==3.11.2
pycodestyle==2.5.0
pylint==2.4.4
python-dotenv==0.10.3
redis==3.3.11
six==1.13.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
